+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/AuthenticationData.java
@@ -30,6 +30,7 @@ import java.util.Map.Entry;
 import org.apache.commons.configuration.Configuration;
 import org.apache.commons.configuration.ConfigurationException;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.reflect.MethodUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.Constant;
@@ -50,6 +51,7 @@ import org.zaproxy.zap.model.Context;
 import org.zaproxy.zap.utils.ZapXmlConfiguration;
 
 public class AuthenticationData extends AutomationData {
+
     public static final String METHOD_HTTP = "http";
     public static final String METHOD_FORM = "form";
     public static final String METHOD_JSON = "json";
@@ -57,6 +59,7 @@ public class AuthenticationData extends AutomationData {
     public static final String METHOD_SCRIPT = "script";
     public static final String METHOD_BROWSER = "browser";
     public static final String METHOD_AUTO = "autodetect";
+    public static final String METHOD_CLIENT = "client";
 
     public static final String PARAM_HOSTNAME = "hostname";
     public static final String PARAM_REALM = "realm";
@@ -69,6 +72,9 @@ public class AuthenticationData extends AutomationData {
     public static final String PARAM_SCRIPT = "script";
     public static final String PARAM_SCRIPT_ENGINE = "scriptEngine";
 
+    // TODO: Plan to change once the core supports dynamic methods better
+    protected static final String CLIENT_SCRIPT_BASED_AUTH_METHOD_CLASSNAME =
+            "org.zaproxy.addon.authhelper.client.ClientScriptBasedAuthenticationMethodType.ClientScriptBasedAuthenticationMethod";
     protected static final String BROWSER_BASED_AUTH_METHOD_CLASSNAME =
             "org.zaproxy.addon.authhelper.BrowserBasedAuthenticationMethodType.BrowserBasedAuthenticationMethod";
 
@@ -76,6 +82,7 @@ public class AuthenticationData extends AutomationData {
     protected static final String FIELD_LOGIN_REQUEST_URL = "loginRequestURL";
 
     private static final String BAD_FIELD_ERROR_MSG = "automation.error.env.auth.field.bad";
+    private static final String PRIVATE_FIELD_SCRIPT = "script";
 
     public static final String VERIFICATION_ELEMENT = "verification";
 
@@ -87,7 +94,8 @@ public class AuthenticationData extends AutomationData {
                     METHOD_JSON,
                     METHOD_SCRIPT,
                     METHOD_BROWSER,
-                    METHOD_AUTO);
+                    METHOD_AUTO,
+                    METHOD_CLIENT);
 
     private String method;
     private Map<String, Object> parameters = new LinkedHashMap<>();
@@ -121,10 +129,29 @@ public class AuthenticationData extends AutomationData {
             JobUtils.addPrivateField(
                     parameters, PARAM_LOGIN_REQUEST_URL, FIELD_LOGIN_REQUEST_URL, jsonAuthMethod);
             JobUtils.addPrivateField(parameters, PARAM_LOGIN_REQUEST_BODY, jsonAuthMethod);
-        } else if (authMethod instanceof ScriptBasedAuthenticationMethod) {
-            ScriptBasedAuthenticationMethod scriptAuthMethod =
-                    (ScriptBasedAuthenticationMethod) authMethod;
-            ScriptWrapper sw = (ScriptWrapper) JobUtils.getPrivateField(scriptAuthMethod, "script");
+        } else if (authMethod != null
+                && authMethod
+                        .getClass()
+                        .getCanonicalName()
+                        .equals(CLIENT_SCRIPT_BASED_AUTH_METHOD_CLASSNAME)) {
+            ScriptWrapper sw =
+                    (ScriptWrapper) JobUtils.getPrivateField(authMethod, PRIVATE_FIELD_SCRIPT);
+            LOGGER.debug("Matched client script class");
+            if (sw != null) {
+                setMethod(METHOD_CLIENT);
+                parameters.put(PARAM_SCRIPT, sw.getFile().getAbsolutePath());
+                parameters.put(PARAM_SCRIPT_ENGINE, sw.getEngineName());
+                @SuppressWarnings("unchecked")
+                Map<String, String> paramValues =
+                        (Map<String, String>) JobUtils.getPrivateField(authMethod, "paramValues");
+                for (Entry<String, String> entry : paramValues.entrySet()) {
+                    parameters.put(entry.getKey(), entry.getValue());
+                }
+            }
+        } else if (authMethod instanceof ScriptBasedAuthenticationMethod scriptAuthMethod) {
+            ScriptWrapper sw =
+                    (ScriptWrapper)
+                            JobUtils.getPrivateField(scriptAuthMethod, PRIVATE_FIELD_SCRIPT);
             if (sw != null) {
                 setMethod(AuthenticationData.METHOD_SCRIPT);
                 parameters.put(PARAM_SCRIPT, sw.getFile().getAbsolutePath());
@@ -291,6 +318,51 @@ public class AuthenticationData extends AutomationData {
                                             .get(AuthenticationData.PARAM_LOGIN_REQUEST_BODY)));
                     context.setAuthenticationMethod(jsonAuthMethod);
                     break;
+                case AuthenticationData.METHOD_CLIENT:
+                    File clientScript =
+                            JobUtils.getFile(
+                                    parameters.getOrDefault(PARAM_SCRIPT, "").toString(),
+                                    env.getPlan());
+                    if (!clientScript.exists() || !clientScript.canRead()) {
+                        progress.error(
+                                Constant.messages.getString(
+                                        "automation.error.env.sessionmgmt.script.bad",
+                                        clientScript.getAbsolutePath()));
+                    } else {
+                        ScriptWrapper sw =
+                                JobUtils.getScriptWrapper(
+                                        clientScript,
+                                        ScriptBasedAuthenticationMethodType.SCRIPT_TYPE_AUTH,
+                                        parameters.getOrDefault(PARAM_SCRIPT_ENGINE, "").toString(),
+                                        progress);
+
+                        AuthenticationMethodType clientScriptType =
+                                extAuth.getAuthenticationMethodTypeForIdentifier(8);
+                        LOGGER.info("Loaded client script auth method type {}.", clientScriptType);
+                        AuthenticationMethod clientScriptMethod =
+                                clientScriptType.createAuthenticationMethod(context.getId());
+
+                        if (sw == null) {
+                            LOGGER.error(
+                                    "Error setting script authentication - failed to find script wrapper");
+                            progress.error(
+                                    Constant.messages.getString(
+                                            "automation.error.env.auth.script.bad",
+                                            clientScript.getAbsolutePath()));
+                        } else {
+                            try {
+                                MethodUtils.invokeMethod(clientScriptMethod, "loadScript", sw);
+                            } catch (Exception e) {
+                                LOGGER.error(e.getMessage(), e);
+                            }
+                            JobUtils.setPrivateField(
+                                    clientScriptMethod, "paramValues", getScriptParameters(env));
+
+                            reloadAuthenticationMethod(clientScriptMethod, progress);
+                            context.setAuthenticationMethod(clientScriptMethod);
+                        }
+                    }
+                    break;
                 case AuthenticationData.METHOD_SCRIPT:
                     File f =
                             JobUtils.getFile(
@@ -308,9 +380,12 @@ public class AuthenticationData extends AutomationData {
                                         ScriptBasedAuthenticationMethodType.SCRIPT_TYPE_AUTH,
                                         parameters.getOrDefault(PARAM_SCRIPT_ENGINE, "").toString(),
                                         progress);
-                        ScriptBasedAuthenticationMethodType scriptType =
+
+                        AuthenticationMethodType scriptType =
                                 new ScriptBasedAuthenticationMethodType();
-                        ScriptBasedAuthenticationMethod scriptMethod =
+                        LOGGER.debug("Loaded script auth method type");
+
+                        AuthenticationMethod scriptMethod =
                                 scriptType.createAuthenticationMethod(context.getId());
 
                         if (sw == null) {
@@ -321,7 +396,11 @@ public class AuthenticationData extends AutomationData {
                                             "automation.error.env.auth.script.bad",
                                             f.getAbsolutePath()));
                         } else {
-                            scriptMethod.loadScript(sw);
+                            try {
+                                MethodUtils.invokeMethod(scriptMethod, "loadScript", sw);
+                            } catch (Exception e) {
+                                LOGGER.error(e.getMessage(), e);
+                            }
                             JobUtils.setPrivateField(
                                     scriptMethod, "paramValues", getScriptParameters(env));