+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Added
+- Added support for Client Script Authentication when used in conjunction with the Ajax Spider add-on.
 
 ## [0.18.0] - 2025-01-27
 ### Changed